# openapi.yaml
openapi: 3.0.1
info:
  title: System Info and Command Runner Plugin
  description: A plugin that allows users to run command line programs after obtaining the system information to accomplish it's goal.
  version: 'v1'
servers:
  - url: http://localhost:5004
paths:
...
  /system/command:
    post:
      operationId: runCommand
      summary: Run a command
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunCommandRequest'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunCommandResponse'
components:
  schemas:
...
    RunCommandRequest:
      type: object
      required:
      - command
      properties:
        command:
          type: string
          description: The command to run.
          required: true
    RunCommandResponse:
      type: object
      properties:
        output:
          type: string
          description: The output of the command.